Fix whitespace and some conventions
Fixed up some whitespace issues when reviewing the tests:
  - Remove whitespace at the end of lines.
  - Indent the instruction in the delay slot.

Replace some usages of exit with 'l.nop 0x1'.  The function exit()
should do the same thing but depends on the stack which is not properly
setup on the .S tests.

Also, be consistent and print 8000000d (good) and exit 0 on success and
baaaaaad (bad) and exit 1 on failure.

Also, when exiting from an exception handler print the vector i.e. 0x700
instead of just exit 1.