Increase tick timer timeout to fix marocchino tests
On Marocchino the exception handlers take a bit longer to run compared
to cappuccino.  This means that the current TT timeout of 30 cycles is
not enough to complete the exception handler.  When marocchino returns
form the exception handler, it can execute 1 instruction then it goes
back to the exception handler.  This means the tests cannot complete.

This change increases the timeout and now the tests are passing.  It
does mean that there will be less timeouts which might cause us to not
hit the same scenarios that we want the timeouts to trigger.  But
monitoring shows we still hit many exceptions.