mmu: Fix page fault count todos
There were a few page faults that were not getting hit as expected, the
issue turned out that we were not running in usermode/supervisor mode
when we expected.

The syscall which was used to switch to supervisor mode no longer works
in newlib (for security reasons).  So after we switch to user mode we
stay there.

The fix is to run user mode tests after a single switch to user mode
at the end of the tests.

Writes to MMU TLB sprs is not enforced to be done in supervisor so we
can do these tests fine.